+ 藥費 4 個 API

+++ src/main/java/tw/com/leadtek/nhiwidget/service/PtAdjustmentFeeService.java
@@ -70,17 +70,18 @@ public class PtAdjustmentFeeService {
             ret += paymentTermsDao.updatePaymentTerms(ptId, params.getFee_no(), params.getFee_name(), params.getNhi_no(), params.getNhi_name(), 
                                                   start_date, end_data, this.Category, 
                                                   params.getHospital_type(), params.getOutpatient_type(), params.getHospitalized_type());
-        
-            if (params.getLst_nhi_no() != null) {
-                paymentTermsDao.deleteExcludeNhiNo(ptId);
-                paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
-            }
-            if (params.getLst_co_nhi_no() != null) {
-                paymentTermsDao.deleteCoexistNhiNo(ptId);
-                paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
-            }
+            if (ret>0) {
+                if (params.getLst_nhi_no() != null) {
+                    paymentTermsDao.deleteExcludeNhiNo(ptId);
+                    paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+                }
+                if (params.getLst_co_nhi_no() != null) {
+                    paymentTermsDao.deleteCoexistNhiNo(ptId);
+                    paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
+                }
 
-            ptAdjustmentFeeDao.update(ptId, params.getCoexist_nhi_no()|0, params.getExclude_nhi_no());
+                ptAdjustmentFeeDao.update(ptId, params.getCoexist_nhi_no()|0, params.getExclude_nhi_no());
+            }
         }
         return ret;
     }
