+++ src/main/java/tw/com/leadtek/nhiwidget/service/PtInpatientFeeService.java
@@ -72,22 +72,23 @@ public class PtInpatientFeeService {
             ret += paymentTermsDao.updatePaymentTerms(ptId, params.getFee_no(), params.getFee_name(), params.getNhi_no(), params.getNhi_name(), 
                                                   start_date, end_data, this.Category, 
                                                   params.getHospital_type(), params.getOutpatient_type(), params.getHospitalized_type());
-        
-            if (params.getLst_nhi_no() != null) {
-                paymentTermsDao.deleteExcludeNhiNo(ptId);
-                paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+            if (ret>0) {
+                if (params.getLst_nhi_no() != null) {
+                    paymentTermsDao.deleteExcludeNhiNo(ptId);
+                    paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+                }
+                if (params.getLst_co_nhi_no() != null) {
+                    paymentTermsDao.deleteCoexistNhiNo(ptId);
+                    paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
+                }
+                if (params.getLst_allow_plan() != null) {
+                    paymentTermsDao.deleteNotAllowPlan(ptId);
+                    paymentTermsDao.addNotAllowPlan(ptId, params.getLst_allow_plan());
+                }
+                ptInpatientFeeDao.update(ptId, params.getMax_inpatient()|0, params.getMax_emergency()|0,  params.getMax_patient_no()|0,
+                                               params.getExclude_nhi_no()|0, params.getNot_allow_plan()|0, params.getCoexist_nhi_no()|0, 
+                                               params.getNo_coexist()|0);
             }
-            if (params.getLst_co_nhi_no() != null) {
-                paymentTermsDao.deleteCoexistNhiNo(ptId);
-                paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
-            }
-            if (params.getLst_allow_plan() != null) {
-                paymentTermsDao.deleteNotAllowPlan(ptId);
-                paymentTermsDao.addNotAllowPlan(ptId, params.getLst_allow_plan());
-            }
-            ptInpatientFeeDao.update(ptId, params.getMax_inpatient()|0, params.getMax_emergency()|0,  params.getMax_patient_no()|0,
-                                           params.getExclude_nhi_no()|0, params.getNot_allow_plan()|0, params.getCoexist_nhi_no()|0, 
-                                           params.getNo_coexist()|0);
         }
         return ret;
     }
