+++ src/main/java/tw/com/leadtek/nhiwidget/service/PtNutritionalFeeService.java
@@ -66,15 +66,16 @@ public class PtNutritionalFeeService {
             ret += paymentTermsDao.updatePaymentTerms(ptId, params.getFee_no(), params.getFee_name(), params.getNhi_no(), params.getNhi_name(), 
                                                   start_date, end_data, this.Category, 
                                                   params.getHospital_type(), params.getOutpatient_type(), params.getHospitalized_type());
-        
-            if (params.getLst_nhi_no() != null) {
-                paymentTermsDao.deleteExcludeNhiNo(ptId);
-                paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+            if (ret>0) {
+                if (params.getLst_nhi_no() != null) {
+                    paymentTermsDao.deleteExcludeNhiNo(ptId);
+                    paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+                }
+                ptNutritionalFeeDao.update(ptId, params.getMax_inpatient()|0, params.getMax_daily()|0, 
+                        params.getEvery_nday()|0, params.getEvery_nday_days()|0, params.getEvery_nday_times()|0, 
+                        params.getOver_nday()|0, params.getOver_nday_days()|0, params.getOver_nday_times()|0, 
+                        params.getExclude_nhi_no()|0);
             }
-            ptNutritionalFeeDao.update(ptId, params.getMax_inpatient()|0, params.getMax_daily()|0, 
-                    params.getEvery_nday()|0, params.getEvery_nday_days()|0, params.getEvery_nday_times()|0, 
-                    params.getOver_nday()|0, params.getOver_nday_days()|0, params.getOver_nday_times()|0, 
-                    params.getExclude_nhi_no()|0);
         }
         return ret;
     }
