+++ src/main/java/tw/com/leadtek/nhiwidget/service/PtTreatmentFeeService.java
@@ -79,28 +79,29 @@ public class PtTreatmentFeeService {
             ret += paymentTermsDao.updatePaymentTerms(ptId, params.getFee_no(), params.getFee_name(), params.getNhi_no(), params.getNhi_name(), 
                                                   start_date, end_data, this.Category, 
                                                   params.getHospital_type(), params.getOutpatient_type(), params.getHospitalized_type());
-        
-            if (params.getLst_nhi_no() != null) {
-                paymentTermsDao.deleteExcludeNhiNo(ptId);
-                paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+            if (ret>0) {
+                if (params.getLst_nhi_no() != null) {
+                    paymentTermsDao.deleteExcludeNhiNo(ptId);
+                    paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+                }
+                if (params.getLst_co_nhi_no() != null) {
+                    paymentTermsDao.deleteCoexistNhiNo(ptId);
+                    paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
+                }
+                if (params.getLst_icd_no() != null) {
+                    paymentTermsDao.deleteIncludeIcdNo(ptId);
+                    paymentTermsDao.addIncludeIcdNo(ptId, params.getLst_icd_no());
+                }
+                if (params.getLst_division() != null) {
+                    paymentTermsDao.deleteLimDivision(ptId);
+                    paymentTermsDao.addLimDivision(ptId, params.getLst_division());
+                }
+                ptTreatmentFeeDao.update(ptId, params.getExclude_nhi_no()|0, params.getCoexist_nhi_no()|0, params.getMax_inpatient()|0, params.getMax_daily()|0, 
+                        params.getEvery_nday()|0, params.getEvery_nday_days()|0, params.getEvery_nday_times()|0, 
+                        params.getPatient_nday()|0, params.getPatient_nday_days()|0, params.getPatient_nday_times()|0, 
+                        params.getMax_patient()|0, params.getInclude_icd_no()|0, params.getMax_month_percentage()|0, 
+                        params.getMax_age()|0, params.getLim_division()|0);
             }
-            if (params.getLst_co_nhi_no() != null) {
-                paymentTermsDao.deleteCoexistNhiNo(ptId);
-                paymentTermsDao.addCoexistNhiNo(ptId, params.getLst_co_nhi_no());
-            }
-            if (params.getLst_icd_no() != null) {
-                paymentTermsDao.deleteIncludeIcdNo(ptId);
-                paymentTermsDao.addIncludeIcdNo(ptId, params.getLst_icd_no());
-            }
-            if (params.getLst_division() != null) {
-                paymentTermsDao.deleteLimDivision(ptId);
-                paymentTermsDao.addLimDivision(ptId, params.getLst_division());
-            }
-            ptTreatmentFeeDao.update(ptId, params.getExclude_nhi_no()|0, params.getCoexist_nhi_no()|0, params.getMax_inpatient()|0, params.getMax_daily()|0, 
-                    params.getEvery_nday()|0, params.getEvery_nday_days()|0, params.getEvery_nday_times()|0, 
-                    params.getPatient_nday()|0, params.getPatient_nday_days()|0, params.getPatient_nday_times()|0, 
-                    params.getMax_patient()|0, params.getInclude_icd_no()|0, params.getMax_month_percentage()|0, 
-                    params.getMax_age()|0, params.getLim_division()|0);
         }
         return ret;
     }
