+++ src/main/java/tw/com/leadtek/nhiwidget/service/PtWardFeeService.java
@@ -65,12 +65,13 @@ public class PtWardFeeService {
             ret += paymentTermsDao.updatePaymentTerms(ptId, params.getFee_no(), params.getFee_name(), params.getNhi_no(), params.getNhi_name(), 
                                                   start_date, end_data, this.Category, 
                                                   params.getHospital_type(), params.getOutpatient_type(), params.getHospitalized_type());
-        
-            if (params.getLst_nhi_no() != null) {
-                paymentTermsDao.deleteExcludeNhiNo(ptId);
-                paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+            if (ret>0) {
+                if (params.getLst_nhi_no() != null) {
+                    paymentTermsDao.deleteExcludeNhiNo(ptId);
+                    paymentTermsDao.addExcludeNhiNo(ptId, params.getLst_nhi_no());
+                }
+                ptWardFeeDao.update(ptId, params.getMin_stay()|0, params.getMax_stay()|0, params.getExclude_nhi_no()|0);
             }
-            ptWardFeeDao.update(ptId, params.getMin_stay()|0, params.getMax_stay()|0, params.getExclude_nhi_no()|0);
         }
         return ret;
     }
